agent_demo/requirements.txt
langchain>=1.0.0,<2.0.0
langchain-openai>=1.0.0,<2.0.0
python-dotenv>=1.0.0,<2.0.0
pytest>=8.0.0,<10.0.0
